==> 01_load-packages.R <==
## R code to process and analyze winter flounder catch data from
## fyke net surveys in the coastal ponds of southern Rhode
## Island. Data were collected by the Rhode Island Division of
## Marine Fisheries from 1999-2024.

## Load packages

library(corrplot)
library(cowplot)
library(dplyr)
library(extrafont)
library(forcats)
library(ggExtra)
library(ggplot2)
library(ggpubr)
library(ggrepel)
library(gtExtras)
library(iml)
library(lubridate)
library(lunar)
library(mousetrap)
library(moments)
library(osmdata)
library(randomForest)
library(readr)
library(readxl)
library(sf)
library(stringr)
library(svglite)
library(tidyr)
library(VSURF)
library(webshot2)
